Improved filter
Wrote an improved Kalman filter which doesn't go unstable when coupled
with control. It does theoretically include most of the dynamics of the
simulator, but it takes too damn long to compute all their derivatives
so it runs super slow.  I then dropped most of the terms and just kept
the gravity due to the sliding masses (in the linearized phi matrix. The
x propagation uses the full dynamics).  Need to experiment more with the
filter. Proportional control about the estimate works ok, but the
problem is that the estimated CoG goes to zero before the real CoG,
which is something that needs to be dealt with.