Almost finished testbed CAD, doing alternate
More or less done with the actuator assemblies, and made an assembly
with the sliding masses, with the z-axis smack-dab in the middle. I want
to make an alternate one where the z-axis is on the side. Committing now
because saving a bunch of copies might mess up my SW references.